cpal
Module platform

Platform-specific items.

This module also contains the implementation of the platform’s dynamically dispatched Host type and its associated Device, Stream and other associated types. These types are useful in the case that users require switching between audio host APIs at runtime.

Structs

AlsaDevice

AlsaDevices

ALSA’s implementation for Devices.

AlsaHost

The default linux, dragonfly, freebsd and netbsd host type.

AlsaStream

Device

The Device implementation associated with the platform’s dynamically dispatched Host type.

Devices

The Devices iterator associated with the platform’s dynamically dispatched Host type.

Host

The platform’s dynamically dispatched Host type.

Stream

The Stream implementation associated with the platform’s dynamically dispatched Host type.

SupportedInputConfigs

The SupportedInputConfigs iterator associated with the platform’s dynamically dispatched Host type.

SupportedOutputConfigs

The SupportedOutputConfigs iterator associated with the platform’s dynamically dispatched Host type.

Enums

DeviceInner

Contains a platform specific Device implementation.

DevicesInner

Contains a platform specific Devices implementation.

HostId

Unique identifier for available hosts on the platform.

HostInner

Contains a platform specific Host implementation.

StreamInner

Contains a platform specific Stream implementation.

Constants

ALL_HOSTS

All hosts supported by CPAL on this platform.

Functions

available_hosts

Produces a list of hosts that are currently available on the system.

default_host

The default host for the current compilation target platform.

host_from_id

Given a unique host identifier, initialise and produce the host if it is available.

Type Aliases

AlsaSupportedInputConfigs

AlsaSupportedOutputConfigs