Struct cpal::platform::Device

pub struct Device(/* private fields */);

The Device implementation associated with the platform’s dynamically dispatched Host type.

Implementations

impl Device

pub fn as_inner(&self) -> &DeviceInner

Returns a reference to the underlying platform specific implementation of this Device.

pub fn as_inner_mut(&mut self) -> &mut DeviceInner

Returns a mutable reference to the underlying platform specific implementation of this Device.

pub fn into_inner(self) -> DeviceInner

Returns the underlying platform specific implementation of this Device.