cpal::platform
Struct Host

pub struct Host(/* private fields */);

The platform’s dynamically dispatched Host type.

An instance of this Host type may represent one of the Hosts available on the platform.

Use this type if you require switching between available hosts at runtime.

This type may be constructed via the host_from_id function. HostIds may be acquired via the ALL_HOSTS const, and the available_hosts function.

Implementations

impl Host

pub fn id(&self) -> HostId

The unique identifier associated with this Host.

pub fn as_inner(&self) -> &HostInner

Returns a reference to the underlying platform specific implementation of this Host.

pub fn as_inner_mut(&mut self) -> &mut HostInner

Returns a mutable reference to the underlying platform specific implementation of this Host.

pub fn into_inner(self) -> HostInner

Returns the underlying platform specific implementation of this Host.