Struct cpal::SupportedFormat

pub struct SupportedFormat {
    pub channels: ChannelCount,
    pub min_sample_rate: SampleRate,
    pub max_sample_rate: SampleRate,
    pub data_type: SampleFormat,
}

Describes a range of supported stream formats.

Fields
channels: ChannelCount min_sample_rate: SampleRate

Minimum value for the samples rate of the supported formats.

max_sample_rate: SampleRate

Maximum value for the samples rate of the supported formats.

data_type: SampleFormat

Type of data expected by the device.

Methods

impl SupportedFormat

pub fn with_max_sample_rate(self) -> Format

Turns this SupportedFormat into a Format corresponding to the maximum samples rate.

pub fn cmp_default_heuristics(&self, other: &Self) -> Ordering

A comparison function which compares two SupportedFormats in terms of their priority of use as a default stream format.

Some backends do not provide a default stream format for their audio devices. In these cases, CPAL attempts to decide on a reasonable default format for the user. To do this we use the "greatest" of all supported stream formats when compared with this method.

Formats are prioritised by the following heuristics:

Channels:

Stereo

Mono

Max available channels

Sample format:

f32

i16

u16

Sample rate:

44100 (cd quality)

Max sample rate