Trait cpal::Sample

pub unsafe trait Sample: Copy + Clone {
    const FORMAT: SampleFormat;

    fn to_f32(&self) -> f32;

    fn to_i16(&self) -> i16;

    fn to_u16(&self) -> u16;

    fn from<S>(_: &S) -> Self
    where
        S: Sample;
}

Trait for containers that contain PCM data.

Associated Constants

const FORMAT: SampleFormat

The SampleFormat corresponding to this data type.

Required methods

fn to_f32(&self) -> f32

Turns the sample into its equivalent as a floating-point.

fn to_i16(&self) -> i16

Converts this sample into a standard i16 sample.

fn to_u16(&self) -> u16

Converts this sample into a standard u16 sample.

fn from<S>(_: &S) -> Self where
S: Sample,

Converts any sample type to this one by calling to_i16, to_u16 or to_f32.