Trait cpal::traits::EventLoopTrait

pub trait EventLoopTrait {
    type Device: DeviceTrait;
    type StreamId: StreamIdTrait;
    fn build_input_stream(
        &self, 
        device: &Self::Device, 
        format: &Format
    ) -> Result<Self::StreamId, BuildStreamError>;

    fn build_output_stream(
        &self, 
        device: &Self::Device, 
        format: &Format
    ) -> Result<Self::StreamId, BuildStreamError>;

    fn play_stream(&self, stream: Self::StreamId) -> Result<(), PlayStreamError>;

    fn pause_stream(
        &self, 
        stream: Self::StreamId
    ) -> Result<(), PauseStreamError>;

    fn destroy_stream(&self, stream: Self::StreamId);

    fn run<F>(&self, callback: F) -> !
    where
        F: FnMut(Self::StreamId, StreamDataResult) + Send;
}

Collection of streams managed together.

Created with the Host::event_loop method.

Associated Types

type Device: DeviceTrait

The Device type yielded by the host.

type StreamId: StreamIdTrait

The type used to uniquely distinguish between streams.

Loading content...
Required methods

fn build_input_stream(
&self,
device: &Self::Device,
format: &Format
) -> Result<Self::StreamId, BuildStreamError>

Creates a new input stream that will run from the given device and with the given format.

On success, returns an identifier for the stream.

Can return an error if the device is no longer valid, or if the input stream format is not supported by the device.

fn build_output_stream(
&self,
device: &Self::Device,
format: &Format
) -> Result<Self::StreamId, BuildStreamError>

Creates a new output stream that will play on the given device and with the given format.

On success, returns an identifier for the stream.

Can return an error if the device is no longer valid, or if the output stream format is not supported by the device.

fn play_stream(&self, stream: Self::StreamId) -> Result<(), PlayStreamError>

Instructs the audio device that it should start playing the stream with the given ID.

Has no effect is the stream was already playing.

Only call this after you have submitted some data, otherwise you may hear some glitches.

Panic

If the stream does not exist, this function can either panic or be a no-op.

fn pause_stream(&self, stream: Self::StreamId) -> Result<(), PauseStreamError>

Instructs the audio device that it should stop playing the stream with the given ID.

Has no effect is the stream was already paused.

If you call play afterwards, the playback will resume where it was.

Panic

If the stream does not exist, this function can either panic or be a no-op.

fn destroy_stream(&self, stream: Self::StreamId)

Destroys an existing stream.

Panic

If the stream does not exist, this function can either panic or be a no-op.

fn run<F>(&self, callback: F) -> ! where
F: FnMut(Self::StreamId, StreamDataResult) + Send,

Takes control of the current thread and begins the stream processing.

Note: Since it takes control of the thread, this method is best called on a separate thread.

Whenever a stream needs to be fed some data, the closure passed as parameter is called. You can call the other methods of EventLoop without getting a deadlock.

Loading content...