Crate cpal

How to use cpal

In order to play a sound, first you need to create a Voice.

// getting the default sound output of the system (can return `None` if nothing is supported)
let endpoint = cpal::get_default_endpoint().unwrap();

// note that the user can at any moment disconnect the device, therefore all operations return
// a `Result` to handle this situation

// getting a format for the PCM
let format = endpoint.get_supported_formats_list().unwrap().next().unwrap();

let mut voice = cpal::Voice::new(&endpoint, &format).unwrap();

Then you must send raw samples to it by calling append_data. You must take the number of channels and samples rate into account when writing the data.

TODO: add example

Important: the append_data function can return a buffer shorter than what you requested. This is the case if the device doesn't have enough space available. It happens very often, this is not some obscure situation that can be ignored.

After you have submitted data for the first time, call play:

voice.play();

The audio device of the user will read the buffer that you sent, and play it. If the audio device reaches the end of the data, it will stop playing. You must continuously fill the buffer by calling append_data repeatedly if you don't want the audio to stop playing.

Structs

Buffer
Represents a buffer that must be filled with audio data.

Endpoint
An opaque type that identifies an end point.

EndpointsIterator
An iterator for the list of formats that are supported by the backend.

Format
Describes a format.

SamplesRate

SupportedFormatsIterator
An iterator that produces a list of formats supported by the endpoint.

Voice
Controls a sound output. A typical application has one Voice for each sound it wants to output.

Enums

ChannelPosition
Possible position of a channel.

CreationError
Error that can happen when creating a Voice.

FormatsEnumerationError
Error that can happen when enumerating the list of supported formats.

SampleFormat
Format that each sample has.

UnknownTypeBuffer
This is the struct that is provided to you by cpal when you want to write samples to a buffer.

Traits

Sample
Trait for containers that contain PCM data.

Functions

get_default_endpoint
Return the default endpoint, or None if no device is available.

get_endpoints_list
Return an iterator to the list of formats that are supported by the system.

Type Definitions

ChannelsCount
Number of channels.