Crate cpal

How to use cpal

In order to play a sound, first you need to create a Voice.

let mut voice = cpal::Voice::new();

Then you must send raw samples to it by calling append_data. This function takes three parameters: the number of channels, the number of samples that must be played per second, and the number of samples that you have available.

You can then fill the buffer with the data.

let mut buffer: cpal::Buffer<f32> = voice.append_data(2, cpal::SamplesRate(44100), 1024);

// filling the buffer with 0s
for e in buffer.iter_mut() {
    *e = 0.0f32;
}

Important: the append_data function can return a buffer shorter than what you requested. This is the case if the device doesn't have enough space available. It happens very often, this is not some obscure situation that can be ignored.

After you have submitted data for the first time, call play:

voice.play();

The audio device of the user will read the buffer that you sent, and play it. If the audio device reaches the end of the data, it will stop playing. You must continuously fill the buffer by calling append_data repeatedly if you don't want the audio to stop playing.

Native format

Each Voice is bound to a specific number of channels, samples rate, and samples format. If you call append_data with values different than these, then cpal will automatically perform a conversion on your data.

If you have the possibility, you should try to match the format of the voice.

Structs

Buffer
Represents a buffer that must be filled with audio data.

SamplesRate

Voice
Controls a sound output. A typical application has one Voice for each sound it wants to output.

Enums

SampleFormat
Format that each sample has.

Traits

Sample
Trait for containers that contain PCM data.

Type Definitions

ChannelsCount
Number of channels.