Module cpal::platform

Platform-specific items.

This module also contains the implementation of the platform's dynamically dispatched Host type and its associated EventLoop, Device, StreamId and other associated types. These types are useful in the case that users require switching between audio host APIs at runtime.

Structs

AlsaDevice

AlsaDevices
ALSA implementation for Devices.

AlsaEventLoop

AlsaHost
The default linux and freebsd host type.

AlsaStreamId

Device
The Device implementation associated with the platform's dynamically dispatched Host type.

Devices
The Devices iterator associated with the platform's dynamically dispatched Host type.

EventLoop
The EventLoop implementation associated with the platform's dynamically dispatched Host type.

Host
The platform's dynamically dispatched Host type.

StreamId
The StreamId implementation associated with the platform's dynamically dispatched Host type.

SupportedInputFormats
The SupportedInputFormats iterator associated with the platform's dynamically dispatched Host type.

SupportedOutputFormats
The SupportedOutputFormats iterator associated with the platform's dynamically dispatched Host type.

Enums

HostId
Unique identifier for available hosts on the platform.

Constants

ALL_HOSTS
All hosts supported by CPAL on this platform.

Functions

available_hosts
Produces a list of hosts that are currently available on the system.

default_host
The default host for the current compilation target platform.

host_from_id
Given a unique host identifier, initialise and produce the host if it is available.

Type Definitions

AlsaSupportedInputFormats

AlsaSupportedOutputFormats