Struct cpal::Buffer

#[must_use]
pub struct Buffer<'a, T: 'a> where T: Sample {
    // some fields omitted
}

Represents a buffer that must be filled with audio data.

You should destroy this object as soon as possible. Data is only committed when it is destroyed.