Struct cpal::Voice

pub struct Voice {
    // some fields omitted
}

Controls a sound output. A typical application has one Voice for each sound it wants to output.

A voice must be periodically filled with new data by calling append_data, or the sound will stop playing.

Each Voice is bound to a specific number of channels, samples rate, and samples format, which can be retreived by calling get_channels, get_samples_rate and get_samples_format. If you call append_data with values different than these, then cpal will automatically perform a conversion on your data.

If you have the possibility, you should try to match the format of the voice.

Methods

impl Voice

fn new(endpoint: &Endpoint, format: &Format) -> Result<Voice, CreationError>

Builds a new channel.

fn format(&self) -> &Format

Returns the format used by the voice.

fn get_channels(&self) -> ChannelsCount

DEPRECATED: use format instead. Returns the number of channels.

You can add data with any number of channels, but matching the voice's native format will lead to better performances.

fn get_samples_rate(&self) -> SamplesRate

DEPRECATED: use format instead. Returns the number of samples that are played per second.

You can add data with any samples rate, but matching the voice's native format will lead to better performances.

fn get_samples_format(&self) -> SampleFormat

DEPRECATED: use format instead. Returns the format of the samples that are accepted by the backend.

You can add data of any format, but matching the voice's native format will lead to better performances.

fn get_period(&self) -> usize

Returns the minimum number of samples that should be put in a buffer before it is processable by the audio output.

If you put less than this value in the buffer, the buffer will not be processed and you risk an underrun.

fn append_data(&mut self, max_samples: usize) -> UnknownTypeBuffer

Adds some PCM data to the voice's buffer.

This function indirectly returns a Buffer object that must be filled with the audio data. The size of the buffer being returned depends on the current state of the backend and can't be known in advance. However it is never greater than max_samples.

You must fill the buffer entirely, so do not set max_samples to a value greater than the amount of data available to you.

Channels are interleaved. For example if you have two channels, you must write the first sample of the first channel, then the first sample of the second channel, then the second sample of the first channel, then the second sample of the second channel, etc.

Panic

Panics if max_samples is 0.

fn play(&mut self)

Sends a command to the audio device that it should start playing.

Has no effect is the voice was already playing.

Only call this after you have submitted some data, otherwise you may hear some glitches.

fn pause(&mut self)

Sends a command to the audio device that it should stop playing.

Has no effect is the voice was already paused.

If you call play afterwards, the playback will resume exactly where it was.

fn get_pending_samples(&self) -> usize

Returns the number of samples in the buffer that are currently being processed by the audio playback backend.

This function is useful to determine how much time it will take to finish playing the current sound.

fn underflowed(&self) -> bool

Returns true if the voice has finished reading all the data you sent to it.