Struct cpal::Voice

pub struct Voice(_);

Controls a sound output. A typical application has one Voice for each sound it wants to output.

A voice must be periodically filled with new data by calling append_data, or the sound will stop playing.

Each Voice is bound to a specific number of channels, samples rate, and samples format, which can be retreived by calling get_channels, get_samples_rate and get_samples_format. If you call append_data with values different than these, then cpal will automatically perform a conversion on your data.

If you have the possibility, you should try to match the format of the voice.

Methods

impl Voice

fn new() -> Voice

Builds a new channel.

fn get_channels(&self) -> ChannelsCount

Returns the number of channels.

You can add data with any number of channels, but matching the voice's native format will lead to better performances.

fn get_samples_rate(&self) -> SamplesRate

Returns the number of samples that are played per second.

You can add data with any samples rate, but matching the voice's native format will lead to better performances.

fn get_samples_format(&self) -> SampleFormat

Returns the format of the samples that are accepted by the backend.

You can add data of any format, but matching the voice's native format will lead to better performances.

fn append_data<'a, T>(&'a mut self, channels: ChannelsCount, samples_rate: SamplesRate, max_elements: usize) -> Buffer<'a, T> where T: Sample + Clone

Adds some PCM data to the voice's buffer.

This function returns a Buffer object that must be filled with the audio data. The size of the buffer being returned depends on the current state of the backend and can't be known in advance. However it is never greater than max_elements.

You must fill the buffer entirely, so do not set max_elements to a value greater than the amount of data available to you.

Channels are interleaved. For example if you have two channels, you must write the first sample of the first channel, then the first sample of the second channel, then the second sample of the first channel, then the second sample of the second channel, etc.

Parameters

channels: number of channels (1 for mono, 2 for stereo, etc.)

samples_rate: number of samples that must be played by second for each channel

max_elements: maximum size of the returned buffer

Panic

Panics if max_elements is 0 or is not a multiple of channels.

fn play(&mut self)

Sends a command to the audio device that it should start playing.

Has no effect is the voice was already playing.

Only call this after you have submitted some data, otherwise you may hear some glitches.

fn pause(&mut self)

Sends a command to the audio device that it should stop playing.

Has no effect is the voice was already paused.

If you call play afterwards, the playback will resume exactly where it was.